Separate out machine checkpointing
