+++ packages/arb-avm-cpp/data_storage/include/data_storage/checkpoint/machinestatesaver.hpp
@@ -19,13 +19,9 @@
 
 #include <avm_values/value.hpp>
 
-struct MachineStateKeys;
 class Transaction;
 struct SaveResults;
 
 SaveResults saveValue(Transaction& transaction, const value& val);
-SaveResults saveMachineState(Transaction& transaction,
-                             const MachineStateKeys& state_data,
-                             uint256_t machineHash);
 
 #endif /* machinestatesaver_hpp */
