+++ packages/arb-avm-cpp/data_storage/include/data_storage/transaction.hpp
@@ -25,6 +25,7 @@
 namespace rocksdb {
 class Transaction;
 class Status;
+struct Slice;
 }  // namespace rocksdb
 
 std::tuple<uint32_t, std::vector<unsigned char>> parseCountAndValue(
