+++ packages/arb-avm-cpp/tests/checkpoint.cpp
@@ -21,6 +21,7 @@
 #include <avm_values/vmValueParser.hpp>
 #include <data_storage/checkpoint/checkpointstorage.hpp>
 #include <data_storage/checkpoint/checkpointutils.hpp>
+#include <data_storage/checkpoint/machine.hpp>
 #include <data_storage/checkpoint/machinestatedeleter.hpp>
 #include <data_storage/checkpoint/machinestatefetcher.hpp>
 #include <data_storage/checkpoint/machinestatesaver.hpp>
